\name{DirNet-package}
\alias{DirNet-package}
\alias{DirNet}
\docType{package}
\title{
\packageTitle{DirNet}
}
\description{
\packageDescription{DirNet}
}
\author{
Riccardo Rastelli <riccardoras@gmail.com>

Juraj Hledik
}
\references{
J. Hledik and R. Rastelli (2018) "A dynamic network model to measure exposure diversification in the Austrian interbank market", \url{https://arxiv.org/abs/1804.01367}
}
\keyword{ package }
