\name{simulated}
\alias{simulated}
\docType{data}

\title{Simulated dataset for the \code{DirNet} package}

\description{
These data have been generated using a homogeneous structure. All edge proportions at all time frames are random samples from a Dirichlet distribution with all of its parameters equal to one.
}

\usage{data(simulated)}

\keyword{datasets}
